+++ src/Computers.h
@@ -28,13 +28,14 @@ class Computers
                 _solver(),
                 _timer_total(0.) ,
                 _timer_solver(0.) ,
-                _timer_pre_proc(0.) 
+                _timer_pre_proc(0.) ,
+                _timer_compute_A(0.)
                 {
                     // make sure that my "grid_model" is ready to be used (for ac and dc)
                     Eigen::Index nb_bus = init_grid_model.total_bus();
-                    CplxVect V = CplxVect::Constant(nb_bus, 1.);
-                    const auto & Vtmp = init_grid_model.get_V();
-                    for(int i = 0; i < Vtmp.size(); ++i) V[i] = Vtmp[i];
+                    CplxVect V = CplxVect::Constant(nb_bus, 1.04);
+                    // const auto & Vtmp = init_grid_model.get_V();
+                    // for(int i = 0; i < Vtmp.size(); ++i) V[i] = Vtmp[i];
                     _grid_model.dc_pf(V, 10, 1e-5);
                     _grid_model.ac_pf(V, 10, 1e-5);
                 };
@@ -56,6 +57,7 @@ class Computers
         double total_time() const {return _timer_total;}
         double solver_time() const {return _timer_solver;}
         double preprocessing_time() const {return _timer_pre_proc;}
+        double amps_computation_time() const {return _timer_compute_A;}
         int nb_solved() const {return _nb_solved;}
 
         // status
@@ -78,6 +80,10 @@ class Computers
         Eigen::Ref<const RealMat > get_flows() const {return _amps_flows;}
         Eigen::Ref<const CplxMat > get_voltages() const {return _voltages;}
         Eigen::Ref<const CplxMat > get_sbuses() const {return _Sbuses;}
+        Eigen::Ref<const RealMat > compute_flows() {
+            compute_flows_from_Vs();
+            return _amps_flows;
+        }
 
     protected:
         template<class T>
@@ -136,8 +142,6 @@ class Computers
             const auto & bus_to = structure_data.get_bus_to();
             const auto & v_yac_ff = structure_data.yac_ff();
             const auto & v_yac_ft = structure_data.yac_ft();
-            const auto & v_yac_tf = structure_data.yac_tf();
-            const auto & v_yac_tt = structure_data.yac_tt();
             Eigen::Index nb_el = structure_data.nb();
             real_type sqrt_3 = sqrt(3.);
 
@@ -152,30 +156,20 @@ class Computers
                 const auto Efrom = _voltages.col(bus_from_me);  // vector (one voltages per step)
                 const auto Eto = _voltages.col(bus_to_me);
 
-                const real_type bus_vn_kv_t = bus_vn_kv(el_id);
-                const RealVect v_f_kv = Efrom.array().abs() * bus_vn_kv_t;
+                const real_type bus_vn_kv_f = bus_vn_kv(bus_from_me);
+                const RealVect v_f_kv = Efrom.array().abs() * bus_vn_kv_f;
 
                 // retrieve physical parameters
                 const cplx_type yac_ff = v_yac_ff(el_id);  // scalar
                 const cplx_type yac_ft = v_yac_ft(el_id);
-                const cplx_type yac_tf = v_yac_tf(el_id);
-                const cplx_type yac_tt = v_yac_tt(el_id);
 
-                // trafo equations
+                // trafo equations (to get the power at the "from" side)
                 CplxVect I_ft =  yac_ff * Efrom + yac_ft * Eto;
-                CplxVect I_tf =  yac_tt * Eto + yac_tf * Efrom;
-
                 I_ft = I_ft.array().conjugate();
-                I_tf = I_tf.array().conjugate();
                 const CplxVect S_ft = Efrom.array() * I_ft.array();
-                const CplxVect S_tf = Eto.array() * I_tf.array();
 
-                const RealVect p_f = S_ft.array().real() * sn_mva;
-                const RealVect q_f = S_ft.array().imag() * sn_mva;
-
-                // now compute the results
-                RealVect res = (p_f.array() * p_f.array() + q_f.array() * q_f.array());
-                res = res.array().sqrt();
+                // now compute the current flow
+                RealVect res = S_ft.array().abs() * sn_mva;
                 res.array() /= sqrt_3 * v_f_kv.array();
                 _amps_flows.col(el_id + lag_id) = res;
             }
@@ -211,5 +205,6 @@ class Computers
         double _timer_total;
         double _timer_solver;
         double _timer_pre_proc;
+        double _timer_compute_A;
 };
 #endif  //COMPUTERS_H